soegemaskine/searchapi/requirements.txt
psycopg[binary,pool]
asyncpg
pgvector
python_dotenv
openai
uvicorn
fastapi
werkzeug